Break model monolith into struct specific files
* Create pkg/model/block_device.go, pkg/model/partition.go, pkg/model/raid_array.go
* Remove RaidArray.Disable() and change RaidArray.Delete() to ultimately use mdadm
  --stop as --replace is used to replace disks within an array.
